# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gorilla Guardians API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: experienceId
          in: query
          schema:
            type: integer
        - name: artisanId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update review status (admin moderation)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # NOTIFICATIONS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, staff, artisan, customer]
        avatar:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        language:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, isActive, createdAt]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: ["integer", "null"]
        experienceId:
          type: ["integer", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: ["string", "null"]
        comment:
          type: string
        images:
          type: array
          items:
            type: string
        isVerifiedPurchase:
          type: boolean
        status:
          type: string
          enum: [pending, approved, rejected]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, userId, rating, comment, images, isVerifiedPurchase, status, createdAt]
    ReviewInput:
      type: object
      required: [rating, comment]
      properties:
        productId:
          type: integer
        experienceId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        comment:
          type: string
        images:
          type: array
          items:
            type: string
    ReviewUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]

    # NOTIFICATION
...